+++ Changelog.md
@@ -41,3 +41,6 @@
   * split the game analysis in 2 scripts
     * video processing
     * report generator
+
+* 20260318-1
+  * opção de gerar um video em formato mp4 com frames analisadas